[issue-25]: Refactor code
Made general structures and functions
