+++ pkg/api/speaker.go
@@ -75,40 +75,20 @@ type DeleteSpeakerForAdminResponse struct {
 
 // API for updating speaker for administration
 type UpdateSpeakerForAdminRequest struct {
-	Name       *string `json:"name"`
-	OnMainPage *bool   `json:"on_main_page"`
-	InFilter   *bool   `json:"in_filter"`
-	Position   *uint64 `json:"position"`
-	Img        *string `json:"img"`
-}
-
-type UpdateSpeakerForAdminResponse struct {
-	ID         uint64    `db:"id"`
-	Name       string    `db:"name"`
-	OnMainPage bool      `db:"on_main_page"`
-	InFilter   bool      `db:"in_filter"`
-	AddedAt    time.Time `db:"added_at"`
-	UpdatedAt  time.Time `db:"updated_at"`
-	Position   uint64    `db:"position"`
-	Img        string    `db:"img"`
+	Name         *string `json:"name"`
+	OnMainPage   *bool   `json:"on_main_page"`
+	InFilter     *bool   `json:"in_filter"`
+	Position     *uint64 `json:"position"`
+	Img          *string `json:"img"`
+	UniversityID *uint64 `json:"university_id"`
 }
 
 // API for adding speaker for administration
 type AddSpeakerForAdminRequest struct {
-	Name       string `json:"name"`
-	OnMainPage bool   `json:"on_main_page"`
-	InFilter   bool   `json:"in_filter"`
-	Position   uint64 `json:"position"`
-	Img        string `json:"img"`
-}
-
-type AddSpeakerForAdminResponse struct {
-	ID         uint64    `db:"id"`
-	Name       string    `db:"name"`
-	OnMainPage bool      `db:"on_main_page"`
-	InFilter   bool      `db:"in_filter"`
-	AddedAt    time.Time `db:"added_at"`
-	UpdatedAt  time.Time `db:"updated_at"`
-	Position   uint64    `db:"position"`
-	Img        string    `db:"img"`
+	Name         string `json:"name"`
+	OnMainPage   bool   `json:"on_main_page"`
+	InFilter     bool   `json:"in_filter"`
+	Position     uint64 `json:"position"`
+	Img          string `json:"img"`
+	UniversityID uint64 `json:"university_id"`
 }
