+++ pkg/app/models/Speaker.go
@@ -3,12 +3,13 @@ package models
 import "time"
 
 type Speaker struct {
-	ID         uint64    `db:"id"`
-	Name       string    `db:"name"`
-	OnMainPage bool      `db:"on_main_page"`
-	InFilter   bool      `db:"in_filter"`
-	AddedAt    time.Time `db:"added_at"`
-	UpdatedAt  time.Time `db:"updated_at"`
-	Position   uint64    `db:"position"`
-	Img        string    `db:"img"`
+	ID           uint64    `db:"id"`
+	Name         string    `db:"name"`
+	OnMainPage   bool      `db:"on_main_page"`
+	InFilter     bool      `db:"in_filter"`
+	AddedAt      time.Time `db:"added_at"`
+	UpdatedAt    time.Time `db:"updated_at"`
+	Position     uint64    `db:"position"`
+	Img          string    `db:"img"`
+	UniversityID uint64    `db:"university_id"`
 }
